# back-end/schema/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Personal trading account API
servers:
  - url: http://localhost:3001/api
paths:
...
  /group/create:
    post:
      tags:
        - group
      summary: Create group
      description: Create group (only with permissions user or higher)
      operationId: createGroup
      requestBody:
        description: Group object that needs to be created
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Group name'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group created'
        '400':
          description: Group already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group already exists'
  /group/{groupId}/delete:
    delete:
      tags:
        - group
      summary: Delete group
      description: Delete group (only with permissions groupOwner or higher)
      operationId: deleteGroup
      parameters:
        - name: groupId
          in: path
          description: ID of group to delete
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group deleted'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group not found'
  /group/{groupId}/update:
    put:
      tags:
        - group
      summary: Update group
      description: Update group (only with permissions groupOwner or higher)
      operationId: updateGroup
      parameters:
        - name: groupId
          in: path
          description: ID of group to update
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      requestBody:
        description: Group object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Group name'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group updated'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group not found'
  /group/{groupId}/invite:
    post:
      tags:
        - group
      summary: Invite user to group
      description: Invite user to group (only with permissions member or higher)
      operationId: inviteUserToGroup
      parameters:
        - name: groupId
          in: path
          description: ID of group to update
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      requestBody:
        description: Group object that needs to be joined
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: _id
                  example: 5f9f1b9b9b9b9b9b9b9b9b9b
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User invited'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group not found'
...
  /group/{groupId}/kick:
    post:
      tags:
        - group
      summary: Kick user from group
      description: Kick user from group (only with permissions groupOwner or higher)
      operationId: kickUserFromGroup
      parameters:
        - name: groupId
          in: path
          description: ID of group to update
          required: true
          schema:
            type: string
            format: _id
            example: 5f9f1b9b9b9b9b9b9b9b9b9b
      requestBody:
        description: Group object that needs to be joined
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: _id
                  example: 5f9f1b9b9b9b9b9b9b9b9b9b
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User kicked'
        '401':
          description: User is not logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User is not logged in'
        '403':
          description: User has no permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'User has no permissions'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Group not found'
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
security:
  - cookieAuth: []
...